pyproject.toml
[project]
name = "finance-analytics-pipeline"
version = "0.1.0"
description = "Data pipeline with Dagster + DuckDB + dbt-duckdb + Great Expectations"
authors = [{ name = "Aleksei Pavlov", email = "insocriscit15@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "dagster==1.9.0",
  "dagster-webserver==1.9.0",
  "great-expectations==0.18.12",
  "google-api-python-client==2.142.0",
  "gspread==6.1.2",
  "python-dotenv==1.0.1",
  "dbt-duckdb>=1.9.6,<2.0.0",
]

[tool.poetry]
packages = [{ include = "orchestration", from = "." }]

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "0.6.4"
black = "24.8.0"
pytest = "8.3.2"
mypy = "^1.18.2"
pre-commit = "^4.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
# Core errors + imports + modernizations + bugbear + simplicities + ruff extras
select = ["E","F","I","UP","B","SIM","C4","RET","PL","RUF"]
ignore = [
  "E501", # Let Black handle line length
  "RUF001" # ignore ambiguous character warnings
]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
warn_unused_ignores = true
disallow_untyped_defs = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "unit: fast, no external I/O",
  "dagster: dagster-focused tests",
  "integration: uses db/files/external processes",
  "e2e: runs full pipeline",
]